+++ artiq/firmware/libboard_artiq/si5324.rs
@@ -179,17 +179,17 @@ fn init() -> Result<()> {
 
     #[cfg(soc_platform = "kasli")]
     {
-        i2c::pca9548_select(BUSNO, 0x70, 0)?;
-        i2c::pca9548_select(BUSNO, 0x71, 1 << 3)?;
+        i2c::switch_select(BUSNO, 0x70, 0)?;
+        i2c::switch_select(BUSNO, 0x71, 1 << 3)?;
     }
     #[cfg(soc_platform = "sayma_amc")]
-    i2c::pca9548_select(BUSNO, 0x70, 1 << 4)?;
+    i2c::switch_select(BUSNO, 0x70, 1 << 4)?;
     #[cfg(soc_platform = "sayma_rtm")]
-    i2c::pca9548_select(BUSNO, 0x77, 1 << 5)?;
+    i2c::switch_select(BUSNO, 0x77, 1 << 5)?;
     #[cfg(soc_platform = "metlino")]
-    i2c::pca9548_select(BUSNO, 0x70, 1 << 4)?;
+    i2c::switch_select(BUSNO, 0x70, 1 << 4)?;
     #[cfg(soc_platform = "kc705")]
-    i2c::pca9548_select(BUSNO, 0x74, 1 << 7)?;
+    i2c::switch_select(BUSNO, 0x74, 1 << 7)?;
 
     if ident()? != 0x0182 {
         return Err("Si5324 does not have expected product number");