+++ RELEASE_NOTES.rst
@@ -35,6 +35,9 @@ Highlights:
 * Removed worker DB warning for writing a dataset that is also in the archive
 * Extended Kasli gateware JSON description with configuration for SPI over DIO.
   See: https://github.com/m-labs/artiq/pull/1800
+* ``PCA9548`` I2C switch class renamed to ``I2CSwitch``, to accomodate support for PCA9547, and
+  possibly other switches in future. Readback has been removed, and now only one channel per 
+  switch is supported. 
 
 Breaking changes: